feat(tracking): Implement basic line-tracking functionality
- Complete line-tracking sensor configuration and initialization
- Add tracking angle calculation algorithm
- Integrate PID control for line-following precision
- Implement core tracking logic in Motor_Proc() function
- Configure multi-sensor tracking system with 5 infrared sensors
- Add error handling and state management for tracking module

Key Changes:
- Created tracking.c with sensor detection logic
- Updated motor control to support tracking mode
- Developed angle calculation and correction mechanisms

Performance:
- Supports smooth and accurate line tracking
- Enables basic autonomous navigation for the robot car

+++ project_3_Car_Learning/Patrol Car/app/encoder/encoder.h
@@ -6,7 +6,7 @@
 void Encoder_Init(void);
 float Encoder_Get_L(void);
 float Encoder_Get_R(void);
-float Encoder_Get_L_Speed(void);
-float Encoder_Get_R_Speed(void);
+int32_t Encoder_Get_L_Speed(void);
+int32_t Encoder_Get_R_Speed(void);
 
 #endif
